:truck: add email-users

+++ camelia/src/models/locale.rs
@@ -3,16 +3,16 @@ use std::path::Path;
 
 use chrono::{NaiveDateTime, Utc};
 use diesel::{delete, insert_into, prelude::*, update};
+use hibiscus::Result;
 use ini::Ini;
 use log::{debug, info, warn};
-use palm::Result;
 use serde::{Deserialize, Serialize};
 
 use super::super::{orm::postgresql::Connection, schema::locales};
 
 #[derive(Queryable, Serialize, Deserialize, Clone)]
 pub struct Item {
-    pub id: i32,
+    pub id: i64,
     pub lang: String,
     pub code: String,
     pub message: String,
@@ -38,11 +38,11 @@ pub trait Dao {
     fn all(&mut self, offset: i64, limit: i64) -> Result<Vec<Item>>;
     fn by_lang(&mut self, lang: &str) -> Result<Vec<Item>>;
     fn by_code(&mut self, code: &str) -> Result<Vec<Item>>;
-    fn by_id(&mut self, id: i32) -> Result<Item>;
+    fn by_id(&mut self, id: i64) -> Result<Item>;
     fn by_lang_and_code(&mut self, lang: &str, code: &str) -> Result<Item>;
-    fn delete(&mut self, id: i32) -> Result<()>;
+    fn delete(&mut self, id: i64) -> Result<()>;
     fn create(&mut self, lang: &str, code: &str, message: &str) -> Result<()>;
-    fn update(&mut self, id: i32, message: &str) -> Result<()>;
+    fn update(&mut self, id: i64, message: &str) -> Result<()>;
 }
 
 fn load_from_ini<P: AsRef<Path>>(
@@ -167,7 +167,7 @@ impl Dao for Connection {
             .load::<Item>(self)?;
         Ok(items)
     }
-    fn by_id(&mut self, id: i32) -> Result<Item> {
+    fn by_id(&mut self, id: i64) -> Result<Item> {
         let it = locales::dsl::locales
             .filter(locales::dsl::id.eq(id))
             .first::<Item>(self)?;
@@ -180,7 +180,7 @@ impl Dao for Connection {
             .first::<Item>(self)?;
         Ok(it)
     }
-    fn update(&mut self, id: i32, message: &str) -> Result<()> {
+    fn update(&mut self, id: i64, message: &str) -> Result<()> {
         let now = Utc::now().naive_utc();
         let it = locales::dsl::locales.filter(locales::dsl::id.eq(id));
         update(it)
@@ -203,7 +203,7 @@ impl Dao for Connection {
             .execute(self)?;
         Ok(())
     }
-    fn delete(&mut self, id: i32) -> Result<()> {
+    fn delete(&mut self, id: i64) -> Result<()> {
         delete(locales::dsl::locales.filter(locales::dsl::id.eq(id))).execute(self)?;
         Ok(())
     }
